fix: infinite-scroll hook 일반적으로 사용할 수 있게 수정
* ref를 훅에서 만들어서 밖에서 사용할 수 있게 함
* 훅 내에서 보여주는 개수 조절할 수 있도록 showNum을 내부에서 정의함
(#19)
